Reorganize the output paths
Change the output paths so that the levels are built in the proper
location for the main application to find them without any further setup
necessary. The main application's output path is bin\{Configuration}\
under the main solution directory, and the levels' output path is the name
of the level under the main application's output path.

Also, remove the hack to redirect the configuration from the Debug folder
to the Release folder. With this setup, it should no longer be necessary:
the Debug build will use the Debug levels, and the Release build will use
the Release levels. However, this means that the Debug and Release
configurations will use a different setup.ini too.